Re-organize into a distributable package
I want to use this in my app :)
