+++ .gitignore
@@ -1,4 +1,6 @@
 *.swp
 *.pyc
 .DS_Store
+VENV/
+*.egg-info/
