+++ setup.py
@@ -1,0 +1,7 @@
+from setuptools import setup
+
+setup(name='chinesepostman',
+      version='0.1',
+      url='https://github.com/supermitch/Chinese-Postman',
+      author='Mitch Leblanc',
+      packages=['chinesepostman'])
